test: move identity map overflow test into existing UoW test file
